initial router structure fixed in frontend
