+++ UnicornCustomer/ClientApp/src/App.tsx
@@ -1,16 +1,12 @@
-import * as React from 'react';
-import { Route } from 'react-router';
-import Layout from './components/Layout';
-import Home from './components/Home';
-import Counter from './components/Counter';
-import FetchData from './components/FetchData';
+import React from "react";
+import Routes from "./routes/Routes";
 
-import './custom.css'
+function App() {
+  return (
+    <div className="App">
+      <Routes />
+    </div>
+  );
+}
 
-export default () => (
-    <Layout>
-        <Route exact path='/' component={Home} />
-        <Route path='/counter' component={Counter} />
-        <Route path='/fetch-data/:startDateIndex?' component={FetchData} />
-    </Layout>
-);
+export default App;
