--- UnicornCustomer/ClientApp/src/components/Layout.tsx
@@ -1,16 +1,0 @@
-import * as React from 'react';
-import { Container } from 'reactstrap';
-import NavMenu from './NavMenu';
-
-export default class Layout extends React.PureComponent<{}, { children?: React.ReactNode }> {
-    public render() {
-        return (
-            <React.Fragment>
-                <NavMenu />
-                <Container>
-                    {this.props.children}
-                </Container>
-            </React.Fragment>
-        );
-    }
-}
