+++ UnicornCustomer/ClientApp/src/modules/PageTest.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const PageTest = () => {
+  return <div>PageTest</div>;
+};
+
+export default PageTest;
