+++ UnicornCustomer/ClientApp/src/modules/PageTestEdit.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const PageTestEdit = () => {
+  return <div>PageTestEdit</div>;
+};
+
+export default PageTestEdit;
